+++ modules/morsecraft.py
@@ -6,28 +6,11 @@ import operator as op
 import numpy as np
 
 import jsonpickle as pickler
-from scripts import modControl as modCon
+
+from .craftmodule import Module as Module
+from .scripts import modControl as modCon
 
 # consider using "uncertainties" package to deal with inaccuracies in coordinate system/simulation
-
-
-class Module:
-    """A module class that contains:
-        position, rotation, connections, type, dimensions, id
-    used within spacecraft"""
-    def __init__(self, mod_id, dimensions=(0.1, 0.1, 0.1), position=(0, 0, 0)):
-        self.cons = [None] * len(dimensions) * 2
-        self.rotation = [1] + [0] * 3
-        self.pos = tuple(np.round(np.array(position), 2))
-        self.type = None
-        self.id = mod_id
-        self.dim = dimensions
-
-    def __str__(self):
-        return self.id
-
-    def __repr__(self):
-        return self.id
 
 
 class Spacecraft:
